add task page as a home hpage
