+++ src/app/feed/page.tsx
@@ -1,0 +1,7 @@
+import FeedPost from "@/components/home/homefeed/FeedPost";
+
+function page() {
+  return <FeedPost />;
+}
+
+export default page;
